requirements.txt
diffusers
transformers
accelerate
opencv-python
numpy
Pillow
matplotlib
tqdm
requests
ollama
fastapi
uvicorn
python-multipart
scikit-image
